+++ packages/awino-ui/src/components/general/Label/Label.tsx
@@ -1,0 +1,41 @@
+import { Tooltip, Typography, TypographyProps, TypographyTypeMap } from '@mui/material';
+import { OverridableComponent } from '@mui/material/OverridableComponent';
+import { styled } from '@mui/material/styles';
+
+import InfoIcon from '@/components/icons/InfoIcon';
+
+interface Props extends Partial<TypographyProps> {
+  tooltip?: string;
+}
+
+const Label = styled(({ children, tooltip, ...restOfProps }: Props) => {
+  return (
+    <Typography {...restOfProps}>
+      {children}
+      {tooltip && (
+        <Tooltip title={tooltip} placement="right">
+          <span className="help">
+            <InfoIcon />
+          </span>
+        </Tooltip>
+      )}
+    </Typography>
+  );
+})(({ theme }) => ({
+  display: 'flex',
+  alignItems: 'center',
+  ...theme.typography.h7,
+  color: theme.palette.text.primary,
+  '.help': {
+    // display: 'inline-block',
+    lineHeight: 0,
+    cursor: 'help',
+    svg: {
+      marginLeft: theme.spacing(2),
+      fontSize: '20px',
+      color: '#03e2e2',
+    },
+  },
+}));
+
+export default Label as OverridableComponent<TypographyTypeMap<Props, 'span'>>;